Add derivationCache and getDerivationCache to BitcoinWalletProvider
derivationCache is useful for speeding up several functions in CAL. However there is no way to instantiate a provider with existing cache. This commit adds getDerivationCache allowing the cache to be exported, and derivationCache to BitcoinWalletProvider, BitcoinJsWalletProvider and BitcoinLedgerProvider allowing a provider to be instantiated with cache from previously created provider.